refactor(portal): promote canvas glue to shared Portal.view.content (#12206)
